+++ Changelog.md
@@ -1,6 +1,30 @@
 ﻿# Changelog World Generation
 
 
+## [0.4.0] - 2023-03-23
+## Added
+- Classes:
+  - TimeManager
+    - New data struct IngameDate
+    - ToString method
+    - Check for correct input
+    - 3 different timeScales
+    - Resume
+    - Pause
+- Changed:
+  - MapGenerator -> GroundGenerator
+  - WaterGenerator
+  - MeshGenerator
+  - GeneratorMain
+
+
+## [0.3.0] - 2023-03-20
+## Added
+- Classes:
+  - TimeManager
+    - Handles ingame time
+
+
 ## [0.2.0] - 2023-03-19
 ## Added
 - Classes:
@@ -9,14 +33,6 @@
     - generate mesh
     - add shader
   - uv-map for mesh generation
-
-
-
-## [0.3.0] - 2023-03-20
-## Added
-- Classes:
-  - TimeManager
-    - Handles ingame time
 
 
 ## [0.1.0] - 2023-03-15